Moved tests package for easier inclusion/exclusion for coverage. Blanket import of test constants. This should be changed, but it makes it work for now. Updated Travis CI script for coveralls support.
