--- gamebench_api_client/tests/__init__.py
@@ -1,3 +1,0 @@
-import os
-
-PARENT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir))
