+++ tests/__init__.py
@@ -1,0 +1,9 @@
+import os
+
+from tests.fixtures.constants import *
+
+PARENT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir))
+
+
+
+
