+++ tests/fixtures/api_samples/auth_sample.json
