+++ tests/fixtures/api_samples/battery_data.json
