+++ tests/fixtures/api_samples/core_frequency.json
