+++ tests/fixtures/api_samples/cpu.json
