+++ tests/fixtures/api_samples/energy.json
