+++ tests/fixtures/api_samples/fps_stability.json
