+++ tests/fixtures/api_samples/gpu_img.json
