+++ tests/fixtures/api_samples/gpu_other.json
