+++ tests/fixtures/api_samples/network.json
