+++ tests/fixtures/api_samples/power.json
