+++ tests/test_unit/test_global_settings.py
