+++ tests/test_unit/test_models/test_dataframes/test_session_detail/test_abstract_session_detail.py
